+++ game/rook.py
@@ -5,7 +5,7 @@ class Rook(StraightMovingPiece):
         super().__init__(color, position)
         
     def symbol(self):
-        return 'R' if self.get_color() == "WHITE" else 'r'
+        return '♖' if self.get_color() == "WHITE" else '♜'
 
     def is_valid_piece_move(self, board, from_pos, to_pos):
         """Verificar si el movimiento es válido para la torre."""